:bug: rename redis to redis_script
